Move collidable line scene
