+++ packages/ag-charts-enterprise/src/features/annotations/scenes/collidableLineScene.ts
